requirements.txt
alembic==1.7.7
altgraph==0.17.2
aniso8601==9.0.1
attrs==21.4.0
auto-py-to-exe==2.18.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.1
Eel==0.12.4
execnet==1.9.0
Flask==2.1.1
Flask-RESTful==0.3.9
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
Mako==1.2.0
MarkupSafe==2.1.1
minio==7.1.5
packaging==21.3
Pillow==9.1.0
pkg_resources==0.0.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pyparsing==3.0.7
pytest==7.1.1
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
whichcraft==0.6.1
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
